Close audit follow-ups: NatRange usize, Rational arithmetic fallibility
Picks up the two remaining "borderline" items from the 2026-05-13
#[expect] audit that PR #598 didn't address. Both turn a silent clippy
suppression into a typed boundary check.

NatRange size is now usize:
- `IndexKind::NatRange { size: usize }` (was `u64`); `IndexDef::step_count`
  drops `*size as u64 -> usize` and its `clippy::cast_possible_truncation`
  suppression.
- `RegistryBuilder::ensure_nat_range_index` takes `usize`; the
  `u64 -> usize` boundary check moves to IR lowering (`ir/lower.rs`),
  where each AST literal carries a span and a graceful
  `GraphcalError::EvalError` is produced if the literal doesn't fit on
  the current target.
- Dynamic registration in `eval_expr/collections.rs` does the same.
- `IndexDef::nat_range_size` keeps the `Option<u64>` accessor signature
  (string-keyed registry name still encodes u64); the lossless
  `usize as u64` cast lives at the accessor.

Rational arithmetic is fallible — no more panics on exponent overflow:
- New `RationalError::Overflow`. `reduce_i64` returns
  `Result<(i32, i32), RationalError>` instead of panicking. Removes the
  `clippy::expect_used` suppression on the GCD-reduction sites.
- `impl Add/Sub/Mul for Rational` now have `Output = Result<Self, _>`;
  `Neg` stays infallible. `Dimension::pow`, `pow_int`, `combine`, and
  the four `Mul/Div` operator impls (owned + `&`) propagate the same
  Result.
- `Registry::resolve_dim_expr / resolve_type_expr / resolve_unit_expr
  / resolve_unit_dimension` return `Result<Option<_>, RationalError>`:
  arithmetic overflow is a hard error, "name not found" stays a soft
  `None`.
- New `GraphcalError::DimensionOverflow { src, span }` (diagnostic code
  `D010`). Every operator call site in `prelude.rs`, `tir/typed.rs`,
  `tir/dim_check/*`, and `ir/lower.rs` wraps `RationalError` at the
  span boundary so users get a proper Miette diagnostic instead of a
  process abort.

Test plan:
- `just lint` — clippy `--all-features` and `--no-default-features`,
  fmt check, doc build, workspace check.
- `just test` — full workspace suite (1,322 tests, 0 failures).

+++ crates/graphcal-compiler/src/tir/dim_check/infer/collections.rs
@@ -98,7 +98,7 @@ pub(super) fn infer_for_comp(
                         InferredType::Scalar(dimension.clone())
                     }
                     crate::registry::types::IndexKind::NatRange { size } => {
-                        InferredType::Fin(NatLinearForm::from_constant(*size))
+                        InferredType::Fin(NatLinearForm::from_constant(*size as u64))
                     }
                 }
             }
